correction on navbar animation
