+++ app/javascript/packs/application.js
@@ -27,7 +27,7 @@ import "bootstrap";
 
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
-import { nav_menu } from "../plugins/_navbar_menu"
+import { nav_menu } from "../plugins/navbar_menu"
 
 document.addEventListener('turbolinks:load', () => {
   // Call your functions here, e.g:
